chore(demo): refactor with latest Cerebral 2 changes
