+++ demos/demo/src/components/Clients/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'cerebral/react'
 import listProps from '../../common/Collection/props/list'
-import translations from '../../common/computed/translations'
+import translations from '../../common/compute/translations'
 
 import Client from '../Client'
