+++ demos/demo/src/components/Date/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {connect} from 'cerebral/react'
-import translations from '../../common/computed/translations'
+import translations from '../../common/compute/translations'
 import {displayDate} from '../../helpers/dateTime'
 
 export default connect(
