+++ demos/demo/src/components/LangSelector/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'cerebral/react'
 import {signal, state} from 'cerebral/tags'
-import translations from '../../common/computed/translations'
+import translations from '../../common/compute/translations'
 import * as LANGS from '../../common/translations'
 
 const LANG_OPTS = Object.keys(LANGS).map(lang => (
