+++ demos/demo/src/components/Project/form.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'cerebral/react'
 import {signal, state} from 'cerebral/tags'
-import translations from '../../common/computed/translations'
+import translations from '../../common/compute/translations'
 
 import Input from './Input'
 import Select from './Select'
