+++ demos/demo/src/components/Textarea/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'cerebral/react'
 import {props, signal, state} from 'cerebral/tags'
-import translations from '../../common/computed/translations'
+import translations from '../../common/compute/translations'
 
 export default connect(
   {
