+++ demos/demo/src/components/Today/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {connect} from 'cerebral/react'
 import {state} from 'cerebral/tags'
 import {displayElapsed, displayTime} from '../../helpers/dateTime'
-import tasksByDay from '../../computed/tasksByDay'
+import tasksByDay from '../../compute/tasksByDay'
 import Date from '../Date'
 
 export default connect(
