+++ demos/demo/src/compute/projectWithDetails.js
