--- demos/demo/src/factories/firebaseItemChanged.js
@@ -1,7 +1,0 @@
-export default function firebaseItemChanged (statePath) {
-  return function mergeItem ({ state, props }) {
-    const { key, value } = props
-    state.set(`${statePath}.${key}`, value)
-    return { key }
-  }
-}
