--- demos/demo/src/factories/firebaseItemRemoved.js
@@ -1,6 +1,0 @@
-export default function firebaseItemRemoved (statePath) {
-  return function removeItem ({ input, state }) {
-    state.unset(`${statePath}.${input.key || input.id}`)
-  }
-}
-
