--- demos/demo/src/modules/tasks/actions/setNow.js
@@ -1,5 +1,0 @@
-import {now} from '../../../helpers/dateTime'
-
-export default function setNow (strings, ...values) {
-  return {now: now()}
-}
