--- demos/demo/src/modules/tasks/actions/startNewTask.js
@@ -1,7 +1,0 @@
-import {now} from '../../../helpers/dateTime'
-
-export default function startNewTask ({state}) {
-  const startedAt = now()
-  state.set(`tasks.$now`, startedAt)
-  state.set(`tasks.$draft.startedAt`, startedAt)
-}
