+++ demos/demo/src/modules/tasks/compute/now.js
@@ -1,0 +1,4 @@
+import {compute} from 'cerebral'
+import {now} from '../../../helpers/dateTime'
+
+export default compute(now)
