--- demos/demo/src/modules/tasks/operators/now.js
@@ -1,5 +1,0 @@
-import {now} from '../../../helpers/dateTime'
-
-export default function (context) {
-  return {value: now()}
-}
