+++ demos/demo/src/modules/tasks/signals/startRunning.js
@@ -1,16 +1,15 @@
 import {set, unset} from 'cerebral/operators'
 import {props, state} from 'cerebral/tags'
-import setNow from '../actions/setNow'
 import paths from '../../../common/Collection/paths'
 import save from '../../../common/Collection/signals/save'
 import updateNow from './updateNow'
+import now from '../compute/now'
 
 const moduleName = 'tasks'
 const {draftPath, errorPath} = paths(moduleName)
 
 export default [
-  setNow,
-  set(state`tasks.$now`, props`now`),
+  set(state`tasks.$now`, now),
   set(state`${draftPath}.key`, 'running'),
   set(state`${draftPath}.startedAt`, state`tasks.$now`),
   unset(state`${draftPath}.endedAt`),
